PricePromoGenix/requirements.txt
flask==2.3.2
pandas==2.0.3
numpy==1.24.3
holidays==0.25
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.0
matplotlib==3.7.1
openai==1.3.0 